feat: change layout and style footer
