--- src/app/shared/components/layouts/Footer/footerList.ts
@@ -1,17 +1,0 @@
-export const footerList = [
-  {
-    id: 1,
-    title: 'Privacy Policy',
-    href: '',
-  },
-  {
-    id: 2,
-    title: 'Term of service',
-    href: '',
-  },
-  {
-    id: 3,
-    title: 'Cooking settings',
-    href: '',
-  },
-];
