--- src/stylesheet/layout/_footer.scss
@@ -1,19 +1,0 @@
-@use '../base/' as *;
-
-.footer {
-  .wrapper {
-    @media screen and (max-width: $md) {
-      justify-content: center;
-      gap: 24px;
-
-      text-align: center;
-    }
-  }
-
-  .logo {
-    img {
-      width: 50%;
-      height: auto;
-    }
-  }
-}
