+++ src/stylesheet/layout/_index.scss
@@ -1,3 +1,2 @@
 @forward './grid';
 @forward './header';
-@forward './footer';
